rebase end_symbol branch, minor changes to make the tests pass
changed '$END' to END in grammar_analysis. Seems to have been added
after the end_symbol branch was created.

moved the end state check to the top of the final loop in lalr_parser.
there was a merge conflict in this file and I guess I did not resolve it
in the correct way